refactor: reorganize and rename files for consistency; remove unused hooks and services
