+++ src/app/api/download/route.ts
@@ -1,4 +1,4 @@
-import dbConnect from '@/lib/dbConnect';
+import dbConnect from '@/lib/db-connect';
 import AppStats from '@/models/app-stats';
 import { NextResponse } from 'next/server';
 import fetch from 'node-fetch';
